Separate the `oep2 report` py.test plugin from the cli tool
