+++ edx_repo_tools/oep2/report/plugin.py
@@ -1,108 +1,15 @@
 """
-``oep2 report``: Check repositories for OEP-2 compliance.
+py.test plugin to support ``oep2 report``. Adds arguments to py.test,
+and creates fixtures for loading git repos.
 """
 
-import logging
 import os.path
-import pkg_resources
-
-import click
-from git.repo.base import Repo
+from git.repo.base import Repo, Head
 from git.refs.remote import RemoteReference
-from git.cmd import Git
 import pytest
 import yaml
+
 from edx_repo_tools.auth import login_github
-
-LOGGER = logging.getLogger(__name__)
-
-
-@click.command()
-@click.option(
-    '-o', '--org',
-    multiple=True,
-    show_default=True,
-    default=[],
-    help="Specify an org to check all of that orgs non-fork "
-         "repositories for OEP-2 compliance",
-)
-@click.option(
-    '-r', '--repo',
-    multiple=True,
-    default=None,
-    help="Specify a repository to check it for OEP-2 compliance",
-)
-@click.option(
-    '--oep',
-    multiple=True, default=None,
-    help="List of OEPs to check for explicit specification of compliance",
-)
-@click.option(
-    '-n', '--num-processes',
-    default=1,
-    help="How many procesess to use while checking repositories",
-)
-@click.option(
-    '--trace/--no-trace', default=False,
-    help="Trace git and github interactions during reporting",
-)
-
-# N.B. We don't use @pass_github here because there isn't a nice way to pass
-# the resulting `hub` object into the pytest tests.
-@click.option(
-    "--checkout-root",
-    default=".oep2-workspace",
-    help="Where to check out repos that are being checked for oep2 compliance",
-)
-@click.option(
-    '--username',
-    help='Specify the user to log in to GitHub with',
-)
-@click.option('--password', help='Password to log in to GitHub with')
-@click.option(
-    '--token',
-    help='Personal access token to log in to GitHub with',
-)
-def cli(org, repo, oep, num_processes, trace, checkout_root, username, password, token):
-    """
-    Command-line interface specification for ``oep2 report``.
-    """
-    args = [
-        '--pyargs', 'edx_repo_tools.oep2.checks',
-        '-c', pkg_resources.resource_filename(__name__, 'oep2-report.ini'),
-        '--checkout-root', checkout_root,
-    ]
-
-    if trace:
-        args.extend(['-s', '-vvv'])
-        Git.GIT_PYTHON_TRACE = True
-    else:
-        args.append('-q')
-
-    if username is not None:
-        args.extend(['--username', username])
-
-    if token is not None:
-        args.extend(['--token', token])
-
-    if password is not None:
-        args.extend(['--password', password])
-
-    if num_processes != 1:
-        args.extend(['-n', num_processes])
-
-    for _org in org:
-        args.extend(['--org', _org])
-
-    for _repo in repo:
-        args.extend(['--repo', _repo])
-
-    for _oep in oep:
-        args.extend(['--oep', _oep])
-
-    click.secho("py.test " + " ".join(args))
-
-    pytest.main(args)
 
 
 def pytest_addoption(parser):
@@ -179,7 +86,7 @@ class Oep2ReportPlugin(object):
         capman = self.config.pluginmanager.getplugin('capturemanager')
 
         capman.suspendcapture(in_=True)
-        self.hub = login_github(
+        hub = login_github(
             self.config.option.username,
             self.config.option.password,
             self.config.option.token,
@@ -188,14 +95,14 @@ class Oep2ReportPlugin(object):
 
         if self.config.option.repo:
             self._repos = [
-                self.hub.repository(*repo.split('/'))
+                hub.repository(*repo.split('/'))
                 for repo in self.config.option.repo
             ]
         elif self.config.option.org:
             self._repos = [
                 repo
                 for org in self.config.option.org
-                for repo in self.hub.organization(org).iter_repos()
+                for repo in hub.organization(org).iter_repos()
                 if not repo.fork
             ]
 
@@ -286,11 +193,14 @@ def git_repo(request, github_repo, branch=None, remote='origin', checkout_root=N
         branch = github_repo.default_branch
 
     head = repo.head
-    target = RemoteReference(repo, 'refs/remotes/{}/{}'.format(remote, branch))
+    remote_branch = RemoteReference(repo, 'refs/remotes/{}/{}'.format(remote, branch))
+    local_branch = Head(repo, 'refs/heads/{}'.format(branch))
 
     try:
-        if head.commit != target.commit:
-            target.checkout()
+        if head.commit != remote_branch.commit:
+            local_branch.commit = remote_branch.commit
+            local_branch.checkout()
+
     except ValueError:
         pytest.xfail("Branch {} is empty".format(branch))
